Move coordinator to module and remove TimeoutError handling
